Removing slashing factor

This [PR aims](https://github.com/keep-network/keep-core/pull/2827#issuecomment-1047732716) to simplify slashing amount calculation. The core idea of how it is calculated [stays the same](https://github.com/keep-network/keep-core/pull/2772#discussion_r776784840) but the difference is in the implementation, ie. removing the slashing factor.

Before:
```
Bytes size for RandomBeacon: 21.644
Gas for submitRelayEntry(): 214122  ·   495215  ·   354669 
```

After:
```
Bytes size for RandomBeacon: 21.606
Gas for submitRelayEntry(): 213826  ·   495062  ·   354444  
```

+++ solidity/random-beacon/contracts/RandomBeaconGovernance.sol
@@ -64,7 +64,7 @@ contract RandomBeaconGovernance is Ownable {
     uint256 public newIneligibleOperatorNotifierReward;
     uint256 public ineligibleOperatorNotifierRewardChangeInitiated;
 
-    uint256 public newRelayEntrySubmissionFailureSlashingAmount;
+    uint96 public newRelayEntrySubmissionFailureSlashingAmount;
     uint256 public relayEntrySubmissionFailureSlashingAmountChangeInitiated;
 
     uint256 public newMaliciousDkgResultSlashingAmount;
@@ -191,7 +191,7 @@ contract RandomBeaconGovernance is Ownable {
         uint256 timestamp
     );
     event RelayEntrySubmissionFailureSlashingAmountUpdated(
-        uint256 relayEntrySubmissionFailureSlashingAmount
+        uint96 relayEntrySubmissionFailureSlashingAmount
     );
 
     event MaliciousDkgResultSlashingAmountUpdateStarted(
@@ -1020,7 +1020,7 @@ contract RandomBeaconGovernance is Ownable {
     /// @param _newRelayEntrySubmissionFailureSlashingAmount New relay entry
     ///        submission failure slashing amount
     function beginRelayEntrySubmissionFailureSlashingAmountUpdate(
-        uint256 _newRelayEntrySubmissionFailureSlashingAmount
+        uint96 _newRelayEntrySubmissionFailureSlashingAmount
     ) external onlyOwner {
         /* solhint-disable not-rely-on-time */
         newRelayEntrySubmissionFailureSlashingAmount = _newRelayEntrySubmissionFailureSlashingAmount;
